refactor(user-profile): extract estimator Cluster 3 coverage helpers (WP2.1d)
Mechanically move the Cluster 3 accuracy & coverage-guidance helpers out of
the `utils/profile_estimator.py` facade into the internal private module
`utils/_profile_estimator/coverage.py` (Deep Refactor Plan v3 Phase 2 — see
docs/user_profile/PROFILE_ESTIMATOR_CLUSTERS.md §2/§3).

Moved verbatim (byte-identical bodies): `_is_lift_filled`, `filled_lift_keys`,
`accuracy_band`, `next_high_impact_lifts`, `cold_start_anchor_lifts`,
`replaced_anchor_lifts` (plus their Issue #17 section-header comment).

Dependency direction (leaf-safe): coverage.py imports only the constants leaf
(ACCURACY_MAJOR_MUSCLE_GROUPS, COLD_START_CANONICAL_COMPOUND,
HIGH_IMPACT_LIFT_PRIORITY, KEY_LIFT_LABELS, KEY_LIFTS) and the core-math
primitives (cold_start_1rm, epley_1rm) — no `strength_calibration` import. The
profile_estimator ⇄ strength_calibration cycle stays held solely by the two
function-local lazy imports in the facade orchestration; unchanged here.

Facade preservation: every moved name is re-exported from the facade via
`from utils._profile_estimator.coverage import (...)  # noqa: F401`, preserving
object identity and the SUPPORTED_ESTIMATOR_EXPORTS surface. The still-in-facade
Clusters 4 (cohort) and 5 (bodymap) plus orchestration continue to reference the
moved names as bare module-level names through this re-export.

pyright baseline: the 2 reportOptionalMemberAccess "get"-on-None diagnostics in
`replaced_anchor_lifts` relocate with the code — facade count 4→2, new
coverage.py count 2 (total unchanged). Surgically updated the committed baseline
for exactly that relocation; all other (CI-environment) entries preserved.

Verification: test_profile_estimator_contract + focused suite (206 passed); full
pytest 1710 passed / 0 failed; blocking flake8 (E9,F63,F7,F82,F811,E711,E712,
F401) clean; pyright baseline gate PASS (0 net-new).

Migration note: none; behavior and public imports preserved.
